+++ src/main/java/com/bashpile/Lexers.java
@@ -22,7 +22,7 @@ import static com.bashpile.shell.BashShell.isWindows;
 /**
  * Helper class for BashpileLexer.
  */
-// TODO find out why bpc bin/bpr.bps runs "type -t commandString:" (trailing colon is incorrect)
+// TODO find out why bpc target/bpr.bps runs "type -t commandString:" (trailing colon is incorrect)
 public class Lexers {
 
     /**