+++ src/main/java/com/bashpile/exceptions/ThrowingFunction.java
@@ -1,0 +1,6 @@
+package com.bashpile.exceptions;
+
+@FunctionalInterface
+public interface ThrowingFunction<E extends Exception> {
+    void apply() throws E;
+}