+++ src/main/java/com/bashpile/exceptions/ThrowingSupplier.java
@@ -1,6 +1,6 @@
 package com.bashpile.exceptions;
 
-/** Used with {@link Exceptions#asUnchecked(ThrowingSupplier)}. */
+/** Used with {@link Exceptions#asUncheckedSupplier(ThrowingSupplier)}. */
 @FunctionalInterface
 public interface ThrowingSupplier<T, E extends Exception> {
     T get() throws E;