+++ src/test/java/com/bashpile/BashpileMainIntegrationTest.java
@@ -25,83 +25,14 @@ public class BashpileMainIntegrationTest extends BashpileTest {
 
     private static final Logger log = LogManager.getLogger(BashpileMainIntegrationTest.class);
 
-    private static boolean bpcDeployed = false;
-
-    private static boolean bprDeployed = false;
-
-    @Test
-    @Timeout(20)
-    @Order(5)
-    public void bpcDeploysSuccessfully() throws IOException {
-        log.info("In bpcDeploysSuccessfully");
-
-        int exitCode = ExecutionResults.GENERIC_FAILURE;
-        int loops = 0;
-        ExecutionResults results = null;
-        while (exitCode != ExecutionResults.SUCCESS && loops++ < 3) {
-            final String command = "bin/bpc bin/bpc.bps";
-            results = runAndJoin(command);
-            log.trace("Output text:\n{}", results.stdout());
-            exitCode = results.exitCode();
-        }
-
-        assertSuccessfulExitCode(results);
-        // ensure generated bpc is executable
-        assertSuccessfulExitCode(runAndJoin("test -x bin/bpc"));
-
-        bpcDeployed = true;
-    }
-
-    @Test
-    @Timeout(20)
-    @Order(6)
-    public void stdlibDeploysSuccessfully() throws IOException {
-        log.info("In stdlibDeploysSuccessfully");
-
-        int exitCode = ExecutionResults.GENERIC_FAILURE;
-        int loops = 0;
-        ExecutionResults results = null;
-        while (exitCode != ExecutionResults.SUCCESS && loops++ < 3) {
-            final String command = "bin/bpc bin/stdlib.bps";
-            results = runAndJoin(command);
-            log.trace("Output text:\n{}", results.stdout());
-            exitCode = results.exitCode();
-        }
-
-        assertSuccessfulExitCode(results);
-    }
-
-    @Test
-    @Timeout(25) // sometimes it takes a while
-    @Order(9)
-    public void bprDeploysSuccessfully() throws IOException {
-        log.info("In bprDeploysSuccessfully");
-        Assumptions.assumeTrue(bpcDeployed);
-
-        // weird, intermittent errors running bpr in Java under WSL like characters getting skipped
-        int exitCode = ExecutionResults.GENERIC_FAILURE;
-        int loops = 0;
-        ExecutionResults results = null;
-        while (exitCode != ExecutionResults.SUCCESS && loops++ < 3) {
-            final String command = "bin/bpc bin/bpr.bps";
-            results = runAndJoin(command);
-            log.trace("Output text:\n{}", results.stdout());
-            exitCode = results.exitCode();
-        }
-
-        assertSuccessfulExitCode(results);
-        bprDeployed = true;
-    }
-
     @Test
     @Timeout(10)
     @Order(10)
     public void bpcTranspiles() throws IOException {
         log.info("In bpcTranspiles test");
-        Assumptions.assumeTrue(bprDeployed);
 
         final String translatedFilename = "src/test/resources/testrigData";
-        final String command = "bin/bpc src/test/resources/testrigData.bps";
+        final String command = "target/bpc src/test/resources/testrigData.bps";
         try {
             final ExecutionResults results = runAndJoin(command);
             log.debug("Output text:\n{}", results.stdout());
@@ -120,11 +51,10 @@ public class BashpileMainIntegrationTest extends BashpileTest {
     @Timeout(20)
     @Order(11)
     public void bprWorks() throws IOException {
-        log.info("In noSubCommandTest");
-        Assumptions.assumeTrue(bprDeployed);
+        log.info("In bprWorks");
 
         // run with our local (not installed) bpr
-        final String command = "bin/bpr src/test/resources/scripts/bprShebang.bps";
+        final String command = "target/bpr src/test/resources/scripts/bprShebang.bps";
         final ExecutionResults results = runAndJoin(command);
         log.debug("Output text:\n{}", results.stdout());
 
@@ -136,10 +66,9 @@ public class BashpileMainIntegrationTest extends BashpileTest {
     @Order(20)
     public void noSubCommandWithNoExtensionTranspiles() throws IOException {
         log.info("In noSubCommandWithNoExtensionTranspiles");
-        Assumptions.assumeTrue(bprDeployed);
 
         final String translatedFilename = "src/test/resources/testrigData2";
-        final String command = "bin/bpc src/test/resources/testrigData2.bps";
+        final String command = "target/bpc src/test/resources/testrigData2.bps";
         Path translatedPath = Path.of(translatedFilename);
         Files.deleteIfExists(translatedPath);
         final ExecutionResults results = runAndJoin(command);
@@ -159,9 +88,8 @@ public class BashpileMainIntegrationTest extends BashpileTest {
     @Order(30)
     public void noSubCommandWithMissingFileFails() throws IOException {
         log.info("In noSubCommandWithMissingFileFails");
-        Assumptions.assumeTrue(bprDeployed);
 
-        final String command = "bin/bpr bin/bpc src/test/resources/testrigData.fileDoesNotExist";
+        final String command = "target/bpc src/test/resources/testrigData.fileDoesNotExist";
         final ExecutionResults results = runAndJoin(command);
         log.debug("Output text:\n{}", results.stdout());
 
@@ -174,10 +102,9 @@ public class BashpileMainIntegrationTest extends BashpileTest {
     @Order(40)
     public void outputFileFlagWithDoubleRunWorks() throws IOException {
         log.info("In outputFileFlagWithDoubleRunWorks");
-        Assumptions.assumeTrue(bprDeployed);
 
         final String translatedFilename = "src/test/resources/testrigData2.example.bps";
-        final String command = "bin/bpc src/test/resources/testrigData2.bps --outputFile " + translatedFilename;
+        final String command = "target/bpc src/test/resources/testrigData2.bps --outputFile " + translatedFilename;
         ExecutionResults results = runAndJoin(command);
         try {
             log.debug("Output text:\n{}", results.stdout());
@@ -199,12 +126,11 @@ public class BashpileMainIntegrationTest extends BashpileTest {
     @Order(50)
     public void bprCreateErrorMessagesPropagate() throws IOException {
         log.info("In bprCreateErrorMessagesPropagate");
-        Assumptions.assumeTrue(bprDeployed);
 
         final String bashpileFilename = "src/test/resources/scripts/overwriteCheck.bps";
         final Path defaultOutputFile = Path.of("src/test/resources/scripts/overwriteCheck");
         assertTrue(Files.exists(defaultOutputFile), "Default output file does not exist");
-        final String command = "bin/bpr " + bashpileFilename;
+        final String command = "target/bpr " + bashpileFilename;
         final ExecutionResults results = runAndJoin(command);
         log.debug("Output text:\n{}", results.stdout());
 
@@ -216,10 +142,9 @@ public class BashpileMainIntegrationTest extends BashpileTest {
     @Order(50)
     public void bprDashCWorks() throws IOException {
         log.info("In bpr -c works");
-        Assumptions.assumeTrue(bprDeployed);
 
         // run with our local (not installed) bpr
-        final String command = "bin/bpr -c \"print('Hello World')\"";
+        final String command = "target/bpr -c \"print('Hello World')\"";
         final ExecutionResults results = runAndJoin(command);
         log.debug("Output text:\n{}", results.stdout());
 
@@ -233,10 +158,9 @@ public class BashpileMainIntegrationTest extends BashpileTest {
     @Order(60)
     public void bpcDashCWithStdinWorks() throws IOException {
         log.info("In bpc -c with stdin works");
-        Assumptions.assumeTrue(bprDeployed);
 
         // run with our local (not installed) bpr
-        final String command = "echo \"print('Hello World')\" | bin/bpc -c";
+        final String command = "echo \"print('Hello World')\" | target/bpc -c";
         final ExecutionResults results = runAndJoin(command);
         log.debug("Output text:\n{}", results.stdout());
 
@@ -251,10 +175,9 @@ public class BashpileMainIntegrationTest extends BashpileTest {
     @Order(70)
     public void bprDashCWithStdinWorks() throws IOException {
         log.info("In bpr -c with stdin works");
-        Assumptions.assumeTrue(bprDeployed);
 
         // run with our local (not installed) bpr
-        final String command = "echo \"print('Hello World')\" | bin/bpr -c";
+        final String command = "echo \"print('Hello World')\" | target/bpr -c";
         final ExecutionResults results = runAndJoin(command);
         log.debug("Output text:\n{}", results.stdout());
 
@@ -268,10 +191,10 @@ public class BashpileMainIntegrationTest extends BashpileTest {
     @Order(80)
     public void bprDashCWithStdinFromDifferentDirectoryWorks() throws IOException {
         log.info("In bpr -c with stdin works");
-        Assumptions.assumeTrue(bprDeployed);
+        Assumptions.assumeTrue(Files.exists(Path.of("../homebrew-bashpile")));
 
         // run with our local (not installed) bpr
-        final String command = "cd ..; echo \"print('Hello World')\" | homebrew-bashpile/bin/bpr -c";
+        final String command = "cd ..; echo \"print('Hello World')\" | homebrew-bashpile/target/bpr -c";
         final ExecutionResults results = runAndJoin(command);
         log.debug("Output text:\n{}", results.stdout());
 
@@ -286,11 +209,11 @@ public class BashpileMainIntegrationTest extends BashpileTest {
     @Order(81)
     public void bprDashCWithStdinFromDifferentDirectoryWithOutputFileWorks() throws IOException {
         log.info("In bpr -c with stdin and outputfile works");
-        Assumptions.assumeTrue(bprDeployed);
+        Assumptions.assumeTrue(Files.exists(Path.of("../homebrew-bashpile")));
 
         // run with our local (not installed) bpr
         final String command =
-                "cd ..; echo \"print('Hello World')\" | homebrew-bashpile/bin/bpr --outputFile command81 -c";
+                "cd ..; echo \"print('Hello World')\" | homebrew-bashpile/target/bpr --outputFile command81 -c";
         final ExecutionResults results = runAndJoin(command);
         log.debug("Output text:\n{}", results.stdout());
 
@@ -305,16 +228,15 @@ public class BashpileMainIntegrationTest extends BashpileTest {
     @Order(81)
     public void bpcDashWithStdinWorks() throws IOException {
         log.info("In bpc - with stdin works");
-        Assumptions.assumeTrue(bprDeployed);
+        Assumptions.assumeTrue(Files.exists(Path.of("../homebrew-bashpile")));
 
         // run with our local (not installed) bpr
         final String command =
-                "cd ..; echo \"print('Hello World')\" | homebrew-bashpile/bin/bpc -";
+                "cd ..; echo \"print('Hello World')\" | homebrew-bashpile/target/bpc -";
         final ExecutionResults results = runAndJoin(command);
         log.debug("Output text:\n{}", results.stdout());
 
         assertSuccessfulExitCode(results);
-        assertTrue(results.stdout().contains("Hello World"));
         assertFalse(Files.exists(Path.of("command.bps")));
         assertFalse(Files.exists(Path.of("../command.bps")));
     }
@@ -322,18 +244,17 @@ public class BashpileMainIntegrationTest extends BashpileTest {
     @Test
     @Timeout(20)
     @Order(82)
-    public void bpcDashWithOutputFileWithStdinWorks() throws IOException {
-        log.info("In bpc - with stdin works");
-        Assumptions.assumeTrue(bprDeployed);
+    public void bpcDashWithOutputFileWorks() throws IOException {
+        log.info("In bpc - with outputFile Works");
+        Assumptions.assumeTrue(Files.exists(Path.of("../homebrew-bashpile")));
 
         // run with our local (not installed) bpr
         final String command =
-                "cd ..; echo \"print('Hello World')\" | homebrew-bashpile/bin/bpc --outputFile dashOutput.bash -";
+                "cd ..; echo \"print('Hello World')\" | homebrew-bashpile/target/bpc --outputFile dashOutput.bash -";
         final ExecutionResults results = runAndJoin(command);
         log.debug("Output text:\n{}", results.stdout());
 
         assertSuccessfulExitCode(results);
-        assertTrue(results.stdout().contains("Hello World"));
         Path dashOutput = Path.of("../dashOutput.bash");
         assertTrue(Files.exists(dashOutput));
 
@@ -346,11 +267,11 @@ public class BashpileMainIntegrationTest extends BashpileTest {
     @Order(84)
     public void bprDashWithStdinWorks() throws IOException {
         log.info("In bpr - with stdin works");
-        Assumptions.assumeTrue(bprDeployed);
+        Assumptions.assumeTrue(Files.exists(Path.of("../homebrew-bashpile")));
 
         // run with our local (not installed) bpr
         final String command =
-                "cd ..; echo \"print('Hello World')\" | homebrew-bashpile/bin/bpr -";
+                "cd ..; echo \"print('Hello World')\" | homebrew-bashpile/target/bpr -";
         final ExecutionResults results = runAndJoin(command);
         log.debug("Output text:\n{}", results.stdout());
 
@@ -364,10 +285,9 @@ public class BashpileMainIntegrationTest extends BashpileTest {
     @Order(90)
     public void bprWithNoArgumentsPrintsHelp() throws IOException {
         log.info("In bpr with no arguments prints help");
-        Assumptions.assumeTrue(bprDeployed);
 
         // run with our local (not installed) bpr
-        final String command = "bin/bpr";
+        final String command = "target/bpr";
         final ExecutionResults results = runAndJoin(command);
         log.debug("Output text:\n{}", results.stdout());
 
@@ -381,9 +301,8 @@ public class BashpileMainIntegrationTest extends BashpileTest {
     @Order(100)
     public void bprDashCFailsGracefullyOnBadCompile() throws IOException {
         log.info("In bpr -c fails gracefully");
-        Assumptions.assumeTrue(bprDeployed);
 
-        final String command = "bin/bpr -c \"# echo\"";
+        final String command = "target/bpr -c \"# echo\"";
         final ExecutionResults results = runAndJoin(command);
 
         assertFailedExitCode(results);