+++ docs/contributor/changelog.md
@@ -34,4 +34,5 @@
 * 0.22.0 - Added increment operator.  Added simpler integer-only translations
 * 0.23.0 - Removed creates statement in favor of createTempFile library call (for tempfile/mktemp)
 * 0.23.1 - Bugfixes (#47 and #48) and refactoring
-* 0.23.2 - Bugfixes and TODO fixes
+* 0.23.2 - Start of optional arguments (only 1), TODO fixes
+* 0.24.0 - Optional arguments and default arguments more generally